fix(utils/ptm): return PageIn event when connecting an intermediate PTE
Before this fix, a PageIn event was not generated when an intermediate page-table entry (level > PT, i.e. PT, PDPT and PML4) was switched from non-present to present.

This fix updates monitor_entry and unmonitor_entry to return a PA if resolved. If monitor_entry returns a PA while in page_in method, it means that the whole chain is paged-in and a PageIn event should be generated.

Note that returning PA from unmonitor_entry is implemented just to mirror the functionality with monitor_entry.